third_party/WebKit/LayoutTests/external/wpt/xhr/overridemimetype-edge-cases.window-expected.txt

This is a testharness.js-based test.
FAIL overrideMimeType() state needs to be reset across requests assert_equals: expected "\ufffd" but got "\ufffd\ufffd"
PASS If charset is not overridden by overrideMimeType() the original continues to be used
FAIL Charset can be overridden by overrideMimeType() with a bogus charset assert_equals: expected "\ufffd" but got "รรฐ"
Harness: the test ran to completion.
